requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
numpy
